Move FILE_FILTER, FILE_TYPES, FILTE_TYPE_MAP from settings object into constants file
