+++ assets/js/Components/Widgets/StatusPill.js
@@ -4,7 +4,7 @@ import SeverityIcon from '../Icons/SeverityIcon'
 import SeverityPotentialIcon from '../Icons/SeverityPotentialIcon'
 import ResolvedIcon from '../Icons/ResolvedIcon'
 import FixedIcon from '../Icons/FixedIcon'
-import { ISSUE_FILTER } from '../../Services/Constants'
+import { FILE_FILTER, ISSUE_FILTER } from '../../Services/Constants'
 
 export default function StatusPill({
   t,
@@ -29,12 +29,12 @@ export default function StatusPill({
           <ResolvedIcon className="color-success icon-md flex-column align-self-center" aria-hidden="true"/>
           <div aria-hidden="true" className="info-pill-text">{t('filter.label.resolution.resolved_single')}</div>
         </div>
-      ) : (issue.status === settings.FILE_FILTER.UNREVIEWED) ? (
+      ) : (issue.status === FILE_FILTER.UNREVIEWED) ? (
         <div aria-label={t('filter.label.resolution.fixed_single')} className="info-pill unreviewed ml-2">
           <SeverityPotentialIcon className="color-potential icon-md flex-column align-self-center" aria-hidden="true"/>
           <div aria-hidden="true" className="info-pill-text">{t('fix.label.status.unreviewed')}</div>
         </div>
-      ) : (issue.status === settings.FILE_FILTER.REVIEWED) ? (
+      ) : (issue.status === FILE_FILTER.REVIEWED) ? (
         <div aria-label={t('filter.label.resolution.resolved_single')} className="info-pill reviewed ml-2">
           <ResolvedIcon className="color-success icon-md flex-column align-self-center" aria-hidden="true"/>
           <div aria-hidden="true" className="info-pill-text">{t('fix.label.status.reviewed')}</div>
